Use correct HTTP method for different SQL statement types

+++ src/main/java/pl/net/was/OpenApiMergeSink.java
@@ -14,6 +14,7 @@
 
 package pl.net.was;
 
+import io.swagger.v3.oas.models.PathItem;
 import io.trino.spi.Page;
 import io.trino.spi.block.Block;
 import io.trino.spi.connector.ConnectorMergeSink;
@@ -44,11 +45,26 @@ public class OpenApiMergeSink
 
     private void updatedPage(Page page, int position)
     {
-        client.putRows(table, page, position);
+        PathItem.HttpMethod method = this.table.getTableHandle().getUpdateMethod();
+        if (method == PathItem.HttpMethod.PUT) {
+            client.putRows(table, page, position);
+        }
+        else if (method == PathItem.HttpMethod.POST) {
+            client.postRows(table, page, position);
+        }
+        else {
+            throw new IllegalArgumentException("Unsupported UPDATE method: " + method);
+        }
     }
 
     private void deletedPage(Block rowIds, int position)
     {
-        client.deleteRows(table, rowIds, position);
+        PathItem.HttpMethod method = this.table.getTableHandle().getDeleteMethod();
+        if (method == PathItem.HttpMethod.DELETE) {
+            client.deleteRows(table, rowIds, position);
+        }
+        else {
+            throw new IllegalArgumentException("Unsupported DELETE method: " + method);
+        }
     }
 }
